Day 3:
Still working with the tip calculator script, but now I've made it an object. Getting aquainted with constructors, the this keyword and arrow functions as methods. My code isn't working though and now I need to work on something else so I have to let it go for the day.